fullscreen/api/document-exit-fullscreen-vs-request-manual.html

<!DOCTYPE html>
<title>Document#exitFullscreen() vs. Element#requestFullscreen()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<div id="parent"><div></div></div>
<script>
async_test(t => {
  const parent = document.getElementById("parent");

  document.onfullscreenchange = t.step_func(() => {
    // We are now in fullscreen, so exiting requires a resize but requesting
    // does not.
    assert_equals(document.fullscreenElement, parent, "fullscreenElement after fullscreenchange event");

    trusted_click(t.step_func(() => {
      // Request fullscreen on another element, to avoid any synchronous
      // short-circuiting on document.fullscreenElement.requestFullscreen(),
      // which used to be in the spec. Also request both before and after the
      // exit. Both requests should be silently ignored due to the exit.
      document.onfullscreenchange = t.step_func(() => {
        // Done, but ensure that there's only one fullscreenchange event.
        document.onfullscreenchange = t.unreached_func("second fullscreenchange event");
        setTimeout(t.step_func_done(), 0);
      });

      const child = parent.firstChild;
      child.requestFullscreen();
      assert_equals(document.fullscreenElement, parent, "fullscreenElement after first requestFullscreen()");
      document.exitFullscreen();
      assert_equals(document.fullscreenElement, parent, "fullscreenElement after exitFullscreen()");
      child.requestFullscreen();
      assert_equals(document.fullscreenElement, parent, "fullscreenElement after second requestFullscreen()");
    }), parent);
  });
  document.onfullscreenerror = t.unreached_func("fullscreenerror event");

  trusted_request(parent);
});
</script>
